requirements.txt
opencv_contrib_python
opencv_python
pyserial
pyinstaller
screeninfo
# cython
# pyobjus